% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_fiatdeposits}
\alias{binance_fiatdeposits}
\title{Get USD Deposits from Binance.US}
\usage{
binance_fiatdeposits(start_time, end_time)
}
\arguments{
\item{start_time}{optional POSIX timestamp}

\item{end_time}{optional POSIX timestamp}
}
\value{
data.table
}
\description{
Get USD Deposits from Binance.US
}
\examples{
\dontrun{
binance_fiatdeposits()
}
}
