% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_fiatwithdrawals}
\alias{binance_fiatwithdrawals}
\title{Get USD withdrawals from Binance.US}
\usage{
binance_fiatwithdrawals(start_time, end_time)
}
\arguments{
\item{start_time}{optional POSIX timestamp}

\item{end_time}{optional POSIX timestamp}
}
\value{
data.table
}
\description{
Get USD withdrawals from Binance.US
}
\examples{
\dontrun{
binance_fiatwithdrawals()
}
}
